♻️ refactor(WritingGoals/Writings) : model > dto로 패키지명 변경 및 그룹화 방식 변경
