+++ src/main/java/com/feelory/feelory_backend/writings/controller/WritingGoalsController.java
@@ -3,7 +3,11 @@ package com.feelory.feelory_backend.writings.controller;
 import com.feelory.feelory_backend.global.api.ApiResponse;
 import com.feelory.feelory_backend.global.api.SuccessCode;
 import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
-import com.feelory.feelory_backend.writings.model.*;
+import com.feelory.feelory_backend.writings.dto.request.WritingGoalCreateRequest;
+import com.feelory.feelory_backend.writings.dto.request.WritingGoalDeleteRequest;
+import com.feelory.feelory_backend.writings.dto.request.WritingGoalListRequest;
+import com.feelory.feelory_backend.writings.dto.request.WritingGoalUpdateRequest;
+import com.feelory.feelory_backend.writings.dto.response.*;
 import com.feelory.feelory_backend.writings.service.WritingGoalsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
