+++ src/main/java/com/feelory/feelory_backend/writings/dto/request/UserWritingDeleteRequest.java
@@ -1,0 +1,12 @@
+package com.feelory.feelory_backend.writings.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
+
+
+@Getter
+@Setter
+public class UserWritingDeleteRequest {
+    @NotNull(message = "id 값은 필수입니다.")
+    private Long id;
+}
