+++ src/main/java/com/feelory/feelory_backend/writings/dto/request/WritingGoalDeleteRequest.java
@@ -1,4 +1,4 @@
-package com.feelory.feelory_backend.writings.model;
+package com.feelory.feelory_backend.writings.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
