+++ src/main/java/com/feelory/feelory_backend/writings/dto/request/WritingGoalListRequest.java
@@ -1,4 +1,4 @@
-package com.feelory.feelory_backend.writings.model;
+package com.feelory.feelory_backend.writings.dto.request;
 
 import lombok.*;
