+++ src/main/java/com/feelory/feelory_backend/writings/dto/request/WritingGoalUpdateRequest.java
@@ -1,16 +1,12 @@
-package com.feelory.feelory_backend.writings.model;
+package com.feelory.feelory_backend.writings.dto.request;
 
-import com.feelory.feelory_backend.global.exception.exceptions.common.InvalidDateFormatException;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
 @Getter
 @Builder
