+++ src/main/java/com/feelory/feelory_backend/writings/dto/response/UserWritingListResponse.java
@@ -1,5 +1,6 @@
-package com.feelory.feelory_backend.writings.model;
+package com.feelory.feelory_backend.writings.dto.response;
 
+import com.feelory.feelory_backend.writings.dto.model.WritingDto;
 import com.feelory.feelory_backend.writings.entity.DailyWordWritings;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
