+++ src/main/java/com/feelory/feelory_backend/writings/dto/response/WritingGoalDetailResponse.java
@@ -1,5 +1,6 @@
-package com.feelory.feelory_backend.writings.model;
+package com.feelory.feelory_backend.writings.dto.response;
 
+import com.feelory.feelory_backend.writings.dto.model.WritingGoalDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,6 +10,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class WritingGoalCreateResponse {
+public class WritingGoalDetailResponse {
     private WritingGoalDto writingGoal;
 }
