+++ src/main/java/com/feelory/feelory_backend/writings/repository/WritingGoalsRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.feelory.feelory_backend.writings.repository;
 
 import com.feelory.feelory_backend.writings.entity.WritingGoals;
-import com.feelory.feelory_backend.writings.model.WritingGoalListSearchDto;
+import com.feelory.feelory_backend.writings.dto.model.WritingGoalListSearchDto;
 import org.springframework.data.domain.Page;
