+++ src/main/java/com/feelory/feelory_backend/writings/service/WritingGoalsService.java
@@ -9,8 +9,14 @@ import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
 import com.feelory.feelory_backend.global.util.ValidationUtil;
 import com.feelory.feelory_backend.users.entity.Users;
 import com.feelory.feelory_backend.users.repository.UsersRepository;
+import com.feelory.feelory_backend.writings.dto.model.WritingGoalDto;
+import com.feelory.feelory_backend.writings.dto.model.WritingGoalListSearchDto;
+import com.feelory.feelory_backend.writings.dto.request.WritingGoalCreateRequest;
+import com.feelory.feelory_backend.writings.dto.request.WritingGoalDeleteRequest;
+import com.feelory.feelory_backend.writings.dto.request.WritingGoalListRequest;
+import com.feelory.feelory_backend.writings.dto.request.WritingGoalUpdateRequest;
+import com.feelory.feelory_backend.writings.dto.response.*;
 import com.feelory.feelory_backend.writings.entity.WritingGoals;
-import com.feelory.feelory_backend.writings.model.*;
 import com.feelory.feelory_backend.writings.repository.WritingGoalsRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
