+++ src/main/java/com/feelory/feelory_backend/writings/service/WritingsService.java
@@ -11,10 +11,12 @@ import com.feelory.feelory_backend.users.entity.Users;
 import com.feelory.feelory_backend.users.repository.UsersRepository;
 import com.feelory.feelory_backend.words.entity.DailyWords;
 import com.feelory.feelory_backend.words.repository.DailyWordsRepository;
+import com.feelory.feelory_backend.writings.dto.model.WritingDto;
+import com.feelory.feelory_backend.writings.dto.request.*;
+import com.feelory.feelory_backend.writings.dto.response.*;
 import com.feelory.feelory_backend.writings.entity.WritingGoals;
-import com.feelory.feelory_backend.writings.model.WritingSearchDto;
+import com.feelory.feelory_backend.writings.dto.model.WritingSearchDto;
 import com.feelory.feelory_backend.writings.entity.DailyWordWritings;
-import com.feelory.feelory_backend.writings.model.*;
 import com.feelory.feelory_backend.writings.repository.DailyWordWritingsRepository;
 import com.feelory.feelory_backend.writings.repository.WritingGoalsRepository;
 import lombok.RequiredArgsConstructor;
